Created print_stack debug tool

+++ include/mstdio.h
@@ -1,7 +1,9 @@
+#include "stdint.h"
+
 #ifndef MSTDIO_H
 #define MSTDIO_H
 void println(const char *str);
 void print(const char *str);
-void printhex(unsigned int number);
+void printhex(uint64_t number);
 void printdec(unsigned int number);
 #endif
